2600 work  Add RIOT and TIA chip stubs -   RIOT has RAM interface working and Register Write Detect   TIA has cpu clock generation and Register Read/Write Detector   6507 generates Phase2 clock   Add RIOT Read/Write ports   Add TIA Read Registers   add controls + missing palette entries   first attempt at HMOVE   add player1 sprite   crude 8k support. Need to move the paging logic to edl rom modules.   Add basic atari mappers   Cart identification improvements + 12k mapper (CBS Ram Plus)   Move unconnected tia registers from write latch   Implement Audio (there are some problems to solve yet)   Add Parker Brothers 8K rom mapper   interpolate audio and fix framerate to ntsc Add all remaining tests to cmake Make contention test self check Put tests in a project folder tests Put EDLSpectrum example in folder examples Add language examples Add SimpleInvaders to cmake Add pin accurate Invaders to cmake Add zx80 to cmake and remove some redundant files Add NES to cmake examples Add vic20 and old style spectrum to cmake Add c64 to cmake

+++ examples/Vic20/1541.c
@@ -8,7 +8,7 @@
 #include <string.h>
 #include <stdint.h>
 
-#include "gui\debugger.h"
+#include "gui/debugger.h"
 
 uint16_t DISK_PinGetPIN_AB();
 uint8_t DISK_PinGetPIN_DB();
@@ -543,7 +543,7 @@ void HandleDiskControllerHardware()		// Every 13(14,15,16) clocks.. advance 1 bi
 				byteAvailable=1;
 				bufBits=0;
 			}
-			DISK_VIA1_PinSetPIN_PA(pBuffer);
+			DISK_VIA1_PinSetPIN_PA((uint8_t)pBuffer);
 		}
 		else
 		{
